RNG seeds are unsigned longs in gsl

+++ include/RandomNumberGenerator.h
@@ -41,7 +41,7 @@ class RandomNumberGenerator
 		RandomNumberGenerator();
 
 		// Use given seed
-		RandomNumberGenerator(int seed);
+		RandomNumberGenerator(unsigned long seed);
 
 		// Free the memory
 		~RandomNumberGenerator();
